+++ packages/@kagal-ed25519-secret/src/index.ts
@@ -14,6 +14,7 @@ export {
 } from './key';
 export {
   type KeyConfig,
+  parseSecretsToKeys,
   parseSecretToKey,
 } from './secret';
 export {